fix(pwa/client): rename middleware to events
