+++ src/client/service-worker/events/lifecycle.js
